fix bad post request, fix auth middleware
