+++ frontend/src/lib/hooks.js
@@ -51,7 +51,7 @@ export const useSignOutUser = () => {
   return () => auth.signOut();
 };
 
-export const useAxiosWithAuth = ({ endpoint, method, body = null }) => {
+export const useAxiosWithAuth = ({ endpoint, method, body = {} }) => {
   const [response, setResponse] = useState(null);
   const [error, setError] = useState('');
   const [status, setStatus] = useState('idle');
@@ -63,30 +63,52 @@ export const useAxiosWithAuth = ({ endpoint, method, body = null }) => {
     const user = auth.currentUser;
     const token = user && (await user.getIdToken());
 
+    // TODO Remove later, used for testing bad token
+    const testToken =
+      'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c';
+
     const config = {
       headers: {
         'Content-Type': 'application/json',
-        Authorization: `Bearer ${token}`,
+        Authorization: `Bearer ${testToken}`,
       },
-      body: JSON.parse(body),
     };
 
-    axios[method](`http://localhost:8080/api/${endpoint}`, config)
-      .then((res) => {
-        setResponse(res.data);
-      })
-      .catch((err) => {
-        setStatus('rejected');
-        setError(err);
-      })
-      .finally(() => {
-        setStatus('resolved');
-      });
+    if (method === 'post') {
+      axios
+        .post(`http://localhost:8080/api/${endpoint}`, body)
+        .then((res) => {
+          setResponse(res.data);
+        })
+        .catch((err) => {
+          setStatus('rejected');
+          setError(err);
+        })
+        .finally(() => {
+          setStatus('resolved');
+        });
+    }
+
+    if (method.toLowerCase === 'get') {
+      axios
+        .get(`http://localhost:8080/api/${endpoint}`, config)
+        .then((res) => {
+          setResponse(res.data);
+        })
+        .catch((err) => {
+          setStatus('rejected');
+          setError(err);
+        })
+        .finally(() => {
+          setStatus('resolved');
+        });
+    }
   }, [body, method, endpoint]);
 
   useEffect(() => {
     fetchData();
-  }, [endpoint, method, body, fetchData]);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
 
   return { response, error, status };
 };
